Lin: reorganization, splitting mbc and campbell_data
