+++ .github/workflows/development-pipeline.yml
@@ -17,7 +17,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.7, 3.8]  # 
+        python-version: [3.6, 3.7, 3.8]  # 
     steps:
       - name: Checkout
         uses: actions/checkout@main
