+++ pyFAST/linearization/__init__.py
@@ -1,3 +1,11 @@
 
-from pyFAST.linearization.tools import MBC_OF
+# NOTE: we make the main functions available here, so that we can change the interface in the future.
+from pyFAST.linearization.tools import getMBCOP, getCDDOP
+from pyFAST.linearization.mbc import fx_mbc3 
 from pyFAST.linearization.campbell import postproCampbell, plotCampbell, plotCampbellDataFile
+from pyFAST.linearization.campbell_data import IdentifyModes
+from pyFAST.linearization.campbell_data import IdentifiedModesDict
+from pyFAST.linearization.campbell_data import printCDDOP
+from pyFAST.linearization.campbell_data import campbellData2TXT
+from pyFAST.linearization.campbell_data import extractShortModeDescription
+from pyFAST.linearization.campbell_data import campbell_diagram_data_oneOP
