+++ pyFAST/linearization/examples/ex1b_OneLinFile_NoRotation.py
@@ -1,6 +1,6 @@
 """ 
 Script to postprocess one linearization file from OpenFAST.
-NOTE: this should not be used if the rotor is turning.
+NOTE: this should not be used if the rotor is turning (multiple lin files would be needed).
 This script would typically be used for a standstill analysis (no rotation),
 or to compute modes of when only isolated degrees of freedom are turned on (and no rotation).
 
@@ -8,7 +8,7 @@ NOTE: should match the script found in the matlab-toolbox
 """
 import os
 import numpy as np
-import pyFAST.linearization.mbc.mbc3 as mbc # TODO will be moved in the future
+import pyFAST.linearization as lin
 
 scriptDir = os.path.dirname(__file__)
 
@@ -16,29 +16,13 @@ scriptDir = os.path.dirname(__file__)
 BladeLen     = 40.04                # Blade length, used to tune relative modal energy [m]
 TowerLen     = 55.59                # Tower length, used to tune relative modal energy [m]
 lin_file     = os.path.join(scriptDir,'../../../data/example_files/Standstill.1.lin') # Linearization file
-#lin_file     = '../../../data/example_files/Standstill_old.1.lin' # Linearization file
-nModesMax    = 10                   # Maximum number of modes to be shown
-nCharMaxDesc = 50                   # Maximum number of characters for description written to screen
 
-## Derived parameters
-lin_files = np.array([lin_file])
-
-# TODO Simplify interface
+# Get Campbell Diagram Data for one Operating Point (CDDOP) given a .lin file
 # Performing MBC (NOTE: not stricly necessary without rotation)
-mbc_data, matData, FAST_linData = mbc.fx_mbc3(lin_files, verbose=False)
-CD = mbc.campbell_diagram_data(mbc_data,BladeLen,TowerLen)
+CDDOP, MBC = lin.getCDDOP([lin_file], BladeLen=BladeLen, TowerLen=TowerLen)
 
 # Outputs to screen
-nModesMax = np.min([len(CD['Modes']),nModesMax])
-Freq = np.array([CD['Modes'][i]['NaturalFreq_Hz'] for i in np.arange(nModesMax)])
-Damp = np.array([CD['Modes'][i]['DampingRatio']   for i in np.arange(nModesMax)])
-print('Mode, NatFreq_[Hz], Damp_Ratio_[-], LogDec._[%], Mode_content_[-]')
-for i in np.arange(nModesMax):
-    Mode = CD['Modes'][i]
-    # Extracting description the best we can
-    Desc = mbc.extractShortModeDescription(Mode)
-    print('{:3d} ,{:12.3f}, {:8.5f}       , {:7.4f},  {:s}'.format(i+1,Mode['NaturalFreq_Hz'],Mode['DampingRatio'],Mode['DampingRatio']*100*2*np.pi, Desc[:min(nCharMaxDesc,len(Desc))]))
-
+Freq,Damp = lin.printCDDOP(CDDOP, nModesMax=10, nCharMaxDesc=50)
 
 if __name__=='__main__':
     pass
